优化Brickyard的代码
